Clarification for company mode completion
New to emacs so not sure if this is was implied or anything. Was using the spacemacs flavour of emacs and wasn't getting any code completion in utop despite seeing it work properly with the merlin-mode. Was able to fix this by adding this to my `~/.emacs`

```elisp`
(add-hook 'after-init-hook 'global-company-mode)
```
Figured might as well try posting in-case it helps anyone else in a similar position.

But as mentioned for some reason this worked out of the box for merlin mode - my (very much a) guess is due to:

```
(require 'company)
```